% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rult-package.R
\docType{package}
\name{rult-package}
\alias{rult}
\alias{rult-package}
\title{rult: A simple interface to RU lexTALE data}
\description{
Use this package to interface with the lextale_ru repo and download data from specific terms or courses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jvcasillas/rult}
  \item Report bugs at \url{https://github.com/jvcasillas/rult/issues}
}

}
\author{
\strong{Maintainer}: Joseph V. Casillas \email{joseph.casillas@rutgers.edu} (\href{https://orcid.org/0000-0001-8735-9910}{ORCID})

}
\keyword{internal}
